refactor: split lexical-editor-wrapper into focused modules
Break the 750-line monolith into a lexical/ subdirectory:

  colored-text-node.ts   - custom TextNode subclass and its span importer
  serialization.ts       - extractNrTag, webvttToLexicalHtml,
                           serializeEditorToHtml (round-trip helpers)
  plugins.tsx            - FocusEmitter, Html, UnmergeableColor,
                           OnChangeHtml lexical plugins
  lexical-toolbar.tsx    - LexicalStaticToolbar
  lexical-editor-wrapper.tsx - the wrapper, initialConfig, and the
                               ContentEditable styling
  index.ts               - barrel exporting the public surface

caption-text-list.tsx now imports from ./lexical. Behavioural tests
move alongside the wrapper. No functional change.
