+++ src/extension/content/feature/editor/components/lexical/index.ts
@@ -1,0 +1,6 @@
+export {
+  LexicalEditorWrapper,
+  type LexicalEditorWrapperProps,
+} from "./lexical-editor-wrapper";
+export { LexicalStaticToolbar } from "./lexical-toolbar";
+export { extractNrTag } from "./serialization";
